Update to implement nextjs project to plangoab
