+++ next.config.js
@@ -1,0 +1,11 @@
+// for transpiling all ESM @fullcalendar/* packages
+// also, for piping fullcalendar thru babel (to learn why, see babel.config.js)
+const withTM = require("next-transpile-modules")(["@fullcalendar"]);
+
+const withImages = require("next-images");
+
+module.exports = withTM(
+  withImages({
+    // any other general next.js settings
+  })
+);
