+++ package.json
@@ -1,25 +1,34 @@
 {
   "name": "plangoab",
-  "version": "0.2.0",
+  "version": "0.1.0",
   "private": true,
-  "homepage": ".",
+  "scripts": {
+    "watch": "next",
+    "build": "next build",
+    "start": "next start",
+    "clean": "rm -rf .next"
+  },
   "dependencies": {
+    "@babel/core": "^7.12.3",
+    "@babel/preset-react": "^7.10.4",
     "@date-io/date-fns": "1.x",
-    "@fullcalendar/interaction": "^5.5.0",
+    "@fullcalendar/common": "^5.4.0",
+    "@fullcalendar/interaction": "^5.4.0",
     "@fullcalendar/list": "^5.5.0",
-    "@fullcalendar/react": "^5.5.0",
-    "@fullcalendar/resource-timeline": "^5.5.0",
+    "@fullcalendar/react": "^5.4.0",
+    "@fullcalendar/resource-timeline": "^5.4.0",
+    "@fullcalendar/timeline": "^5.4.0",
     "@material-ui/core": "^4.11.2",
     "@material-ui/pickers": "^3.2.10",
     "@reduxjs/toolkit": "^1.5.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
+    "babel-plugin-transform-require-ignore": "^0.1.1",
     "date-fns": "^2.16.1",
+    "next": "^10.2.0",
+    "next-images": "^1.7.0",
+    "next-transpile-modules": "^4.1.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-hook-form": "^6.15.1",
@@ -27,17 +36,8 @@
     "react-scripts": "4.0.1",
     "redux-persist": "^6.0.0",
     "styled-components": "^5.2.1",
-    "typescript": "^4.0.3",
     "uuid": "^8.3.2",
     "web-vitals": "^0.2.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "graph:dot": "yarn depcruise --include-only '^src' --output-type dot ./src --output-to dependencygraph.dot",
-    "graph:svg": "dot dependencygraph.dot -T svg > dependencygraph.svg"
   },
   "eslintConfig": {
     "extends": [
@@ -63,11 +63,16 @@
   },
   "devDependencies": {
     "@types/date-fns": "^2.6.0",
+    "@types/jest": "^26.0.15",
+    "@types/node": "^15.0.2",
+    "@types/react": "^17.0.5",
+    "@types/react-dom": "^16.9.8",
     "@types/react-redux": "^7.1.16",
     "@types/styled-components": "^5.1.7",
     "@types/uuid": "^8.3.0",
     "@types/webpack-env": "^1.16.0",
     "dependency-cruiser": "^9.23.1",
-    "eslint-plugin-react-hooks": "^4.2.0"
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "typescript": "^4.2.4"
   }
 }
