--- plangoab/next.config.js
@@ -1,10 +1,0 @@
-
-// for transpiling all ESM @fullcalendar/* packages
-// also, for piping fullcalendar thru babel (to learn why, see babel.config.js)
-const withTM = require('next-transpile-modules')([
-  '@fullcalendar'
-])
-
-module.exports = withTM({
-  // any other general next.js settings
-})
