--- plangoab/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "plangoab",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "watch": "next",
-    "build": "next build",
-    "start": "next start",
-    "clean": "rm -rf .next"
-  },
-  "dependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/preset-react": "^7.10.4",
-    "@fullcalendar/common": "^5.4.0",
-    "@fullcalendar/interaction": "^5.4.0",
-    "@fullcalendar/react": "^5.4.0",
-    "@fullcalendar/resource-timeline": "^5.4.0",
-    "@fullcalendar/timeline": "^5.4.0",
-    "babel-plugin-transform-require-ignore": "^0.1.1",
-    "next": "^10.2.0",
-    "next-transpile-modules": "^4.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
-  }
-}
