--- plangoab/pages/_app.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-import "@fullcalendar/common/main.css";
-import "@fullcalendar/timeline/main.css";
-import "@fullcalendar/resource-timeline/main.css";
-
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
