--- src/index.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-import { ReactReduxProvider } from './providers/ReactReduxProvider'
-
-const render = () => {
-  ReactDOM.render(
-    <React.StrictMode>
-      <ReactReduxProvider>
-        <App />
-      </ReactReduxProvider>
-    </React.StrictMode>,
-    document.getElementById("root")
-  );
-};
-
-render();
-
-// for redux
-if (process.env.NODE_ENV === "development" && module.hot) {
-  module.hot.accept("./App", render);
-}
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.warn))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
