+++ src/pages/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { CalendarPage } from "../components/pages/CalendarPage";
+
+const Index = () => <CalendarPage />;
+
+export default Index;
